Make sourcehut namespace configurable
Replace hardcoded ~pmos namespace with an actually configurable
namespace. The namespace shows up in sr.ht links, and while it works to
put any value there, as long as the ID is right, it leads to confusion
to have it not pointing to the user that is running the test.

I have named the variable "namespace" and included the ~ character,
because sr.ht is getting support for something like organizations at
some point. Maybe then setting another character is required.