SrHtApi.php: fix repoRoot
This piece of code should have check if a repository exist, so we don't
pass non-existing repository URLs. But so far, the check was broken, so
it never added any repositories. Causing to always build all packages
from scratch, even if we had built them already and they did not change.