SrHtApi.php: fix uploading subpackages
Do not hardcode the package name in the glob of apks to be uploaded.
Instead, upload everything in $work/packages/$arch/*.apk.

This should fix uploading subpackages, which do not start with the
package name, such as the "musl-dev-aarch64" subpackage of
"musl-aarch64". The currently broken behavior can be seen in [1], and
causes gcc-aarch64 and similar package builds to fail, because they miss
musl-dev-aarch64 [2].

Using *.apk is safe, since we started using
'pmbootstrap build --no-depends' recently. So we are sure that we only
build the package we are intending to build, and therefore only apks for
that will be produced and catched by that glob.

[1]: https://builds.sr.ht/~martijnbraam/job/55256
[2]: https://builds.sr.ht/~martijnbraam/job/55885