Initial commit for python version
We had a PHP based version that is about 90% feature complete, but had
some hard to resolve bugs and no test cases. Long story short, we made
the tough decision of starting over with a Python 3 version and with
focus on testing from the beginning. Using the same programming language
as pmbootstrap decreases friction for development, for writing tests,
and increases the bus factor.

Related: https://postmarketos.org/blog/2019/06/23/two-years/#sourcehut-srht