Update config code and various other changes
move:
	bpo/helpers/config.py -> bpo/config/args.py
	bpo/helpers/configfile.py -> bpo/config/tokens.py

Re-enable the token logic again, remove the old args code from it, and
make it work more like the new config.py/args.py code. Provide a test
token file, that can be used with curl.sh.

The config file only holds the tokens and nothing else anymore. This is
good enough for this little build management program; for a production
environment, or if we want to have certain args always set for other
reasons, we can simply set up a shell script wrapper for bpo.py, and
avoid implementing complicated logic to mix up args and config file
contents.

I'm planning to move bpo/helpers/constants.py to bpo/config/const.py.

config.py: it would be nice to set the default automatically, and to
automatically mention the default in the description.