Use sqlalchemy
After researching, how well python3 + sqlite3 works with threads (since
we probably want to use threads in the web server), I came to the
conclusion that we might as well switch to sqlalchemy. Then we could
also easily swap out the database backend, if we want to do that at some
point.