Get most of repo_missing callback working
Yesterday I dug into this again, after a 1.5 month break that was
required to write the two years of postmarketOS blog post etc. This is
still early development, with obviously not a one-change-for-one-commit
style git log. So changes are all over the place, but here are the most
important things.

repo_missing:
* Make repo_missing callback insert/update missing packages in db
* Require X-BPO-Arch and X-BPO-Push-Id headers for the repo_missing
  callback
* Make an unique index across pkgname-arch in the packages table
* Add sh test script for repo_missing callback

running jobs:
* everything is sent to submit.py with environment variables now, so
  we can do it more easily from bpo (see related changes in bpo-python
  branch on pmaports.git)
* when running locally (running on sourcehut is not implemented yet),
  properly write the test token to a file and use it with submit.py
* when running locally, properly export the host
* rename jobs for consistent naming

In theory it should be possible now to trigger the push_hook, which
runs "pmbootstrap repo_missing" and then calls the repo_missing callback
with submit.py. However, pmbootstrap is hanging there currently in an
infinite loop for some reason, and I have not investigated it yet.
Saving that for next time :) (I've reorganized a few things and should
now be able to consistently spend a few hours every week on bpo, so this
is finally making progress again!)