Figure out the next package to build and build it
Regarding the new status, build_id columns in the package table:
I thought about having a queue table or not, and came to the conclusion
that having a queue table would only increase complexity. We would need
to update the queue table whenever the repo_missing callback fires.
Let's keep the package build status in the packages table instead, and
not rely on a queue table, but rather find the next package to be built
on-demand.

A nice side-effect is, that this will allow building multiple packages
in parallel, in case we want to do that in the future. Because the
queue will change depending on which package finishes building first, so
there's no point keeping a queue around in that case at all.