tools: add abuild-sign-noinclude, apk-tools-static
Add tools needed to generate an APKINDEX on the bpo server. Just like
the PHP version did it, the plan is to generate the index on the
server, then run a job on another server (sourcehut) to sign the index.
That way, we don't need to download the whole repository to the
sourcehut server for indexing!

The apks were copied from the postmarketOS and Alpine repositories. In
the future we could add a gitlab-ci job, that makes sure that we keep
these up-to-date. Note that they are just used for generating the
index, not for transfering files over the network or the likes, so there
is a much lower attack surface.

These are x86_64 binaries, using bpo on other architectures is
out-of-scope currently (would be easy to add that if needed in the
future though).