Generate WIP repo indexing key during init
I shortly thought about whether we need to sign the WIP repository
APKINDEX at all, considering that we will be downloading it with HTTPS
from where the next package built job runs. But apk expects indexes to
be signed, so simply signing it locally is the easiest way.

I'm planning to provide the pub key to the build job as part of the
build script that we submit to the building backend (sourcehut, local).