jobs: always run submit.py with exec
submit.py is blocking, so the shell interpreter waits until the command
completes. If we leave out the "exec", then the shell interpreter will
reload the file after submit.py completed and execute "the next line".

When running locally, in most cases we will already have changed the
current_task.sh script when submit.py completes. So the shell parser
continues to read garbage from the changed file at a random location
and the job fails.

I thought about giving the temporary task file a random name instead, to
avoid this collision. But that makes it less convenient for me, because
I can't simply have the file open in my editor and reload it after
running a test.