job_services/local: copy WIP repo over local pkgs
Instead of starting a second webserver, that hosts the WIP repository
packages, just copy them over the locally built packages. This
simplifies the development and test setup. The locally built packages
are empty anyway (zapped right before) as we are building packages for
the repository, so this is fine.

In order to do this, the temp dir needs to be set up before copying
the WIP repo contents, so we have the symlink to the pmaports ready. So
refactor the setup code of the local job service to set up that temp dir
during the setup shell script, instead of doing it later with python
code.