test: remove leftovers
Remove old shell test scripts, that require manual changes to
pmaports.git. We have the fully automated pytest-based testsuite now.

test/test_apkindex.sh served as example to use the tools, for creating
an APKINDEX and signing it. But that functionallity is now implemented
in the bpo server.

test/common.py was never used.