Disable threading in flask server (werkzeug)
Let's make the server code less complex and avoid some race conditions.

My reasoning is, that it does not matter much, if an API all coming
from gitlab or sourcehut is hanging for a few seconds. Regarding the
user-facing website, I'm planning to just throw a bunch of HTML files
from the bpo server to the disk. Then have a regular, of course
threaded, httpd display these pages. So the web UI would not be affected
by this design decision.