bpo.ui.log_and_update: new function
Add a new log_and_update() function, which, as the name implies, logs
to the log table and updates the HTML output right afterwards. We need
some kind of trigger to when the HTML output should be updated, and it
should not be after each and every database change. But having one log
entry after each important db change, and then rewriting, seems like a
good amount.