bpo.main(): use log_and_update()
Trigger the first write to the HTML output by inserting a new log
message, so we have a nice message in the UI that the server was just
restarted. In the future, we could also make it show the git commit
of the bpo code, that the server is running.