Remove Push and Commit tables
In the next patch, I will add a --auto-get-repo-missing argument to
bpo.py. In order to make this happen, the get_repo_missing job must not
require a push object anymore. So I realized, that the push, and even
the commit tables are not that useful. They could be used to display the
commits associated with a certain push in the UI, but then again, we can
also read this out of the payload from the log table. So let's unclutter
the database a bit and remove this.