jobs: generate and return job id
Do not start a dedicated thread for each job in bpo.helpers.job.run, but
do it in bpo.job_services.local instead. Generate a random number as job
id in the local job service and return it. With sourcehut, we will block
until the sourcehut server accepts the job and returns the job id.

Log the id for now, the plan is to store it in the database.