job_services/local: set BPO_JOB_{ID,NAME}
pmaports.git's submit.py is adjusted accordingly, so both variables get
sent to the bpo server. This is necessary for handling job failures.
Strictly speaking, we would not need to send the job name if we already
have the ID, but it makes the code less complex as we can just switch
to the proper job type handling by the name.