jobs: refactor package status update after restart
Move most of the logic from the local job service to the common code in
bpo.helpers.job.update_package_status(). Replace the local job services'
update_package_status_after_restart() function with a small get_status()
function.

While this always returns "failed" for the local job, the idea is that
this will really check the status with the sourcehut api for the
sourcehut backend. So we can properly check if jobs are still running
after restarting the bpo server, and if they are not, mark them as built
or failed and start the next job.

Get rid of the "after_restart" bit in the function name, so we can also
call it on other occasions.