job services: start work on sourcehut integration
Add the commandline arguments, a shell script example, require the token
to be set in the tokens file and adjust the README.

For anyone wondering why there had been so many commits before starting
work on sourcehut integration, although this is all about moving the
binary repository package building to sourcehut: the complex parts of
this bpo server code have nothing to do with sourcehut. Therefore I had
put the focus on the local backend and testing so far.