bpo/jobs/build_package: add workaround for gcc*-*
Building cross compilers the way we do it in postmarketOS is currently
broken with the latest apk-tools and abuild, due to this bug:
https://gitlab.alpinelinux.org/alpine/apk-tools/issues/10649

This is fixed in the master branch of abuild, but not yet in the latest
release. Add a config option to disable the --strict argument for
"pmbootstrap build" for certain packages, and make it generic in case we
need this workaround for other packages in the future.