job_services: remove setup_task
Having a dedicated function for generating the setup_task only makes the
sourcehut code more complicated. Because not all of the job's variables
are available when setup_task gets called. So move the code for the
local job service to the LocalJobServiceThread and remove setup_task
from all job service classes.