bpo.repo.wip.clean: rm apks without origin in db
Extend bpo.repo.wip.clean() to not only delete apks that are in the
final repo, but also to delete apks that are not in the database
anymore (with the same version). This could happen, if a package was
updated or removed in a pmaports.git push, which had already been built
for the WIP repo, but was not published at this point. Add a test, that
goes through all code paths of clean().