bpo.repo.symlink.link_to_all_packages: fix subpkgs
Replace the old algorithm of finding each package from the database in
either the WIP or final repo and creating a symlink to the symlink
repo. This did not take subpackages into account.

Instead, use the find-in-WIP-or-final-repo code as sanity check. Clean
outdated packages in the WIP repo and then create a symlink for all
remaining apks in the WIP repo to the symlink repo. This also catches
subpackages. Finally, iterate over the apks in the final repo dir, and
create a symlink to each apk, where the origin is found in the database
(with the same version, so we know it is up-to-date). Add test for all
codepaths.

Fixes: #46