cb/get_repo_missing: delete some packages from db
First attempt to delete packages from the database, that have been
removed from pmaports.git. This improves the situation, but does not
entirely resolve the problem.

We should be able to delete packages that never made it into the final
repository with this code, but once they are there, they will be kept
forever. In order to also delete packages that made it into the final
repository, but which are no longer in the pmaports, we need a list of
all packages in pmaports.git for the given arch and branch. Right now
we only have the packages that need to be built.

In other words, the job calling "pmbootstrap repo_missing" needs to send
more information back to the bpo server.

Related: #38