callback/get_repo_missing: commit before ui.log()
Fix the "database is locked" error by committing all changes done in the
current database sessions, before running bpo.ui.log() (and therefore
opening another session). Instead of doing the commit at the end of
several helper functions in the file, do the commit at the end of each
function, to make it less likely that this mistake is made again.