callback/build_package: update job_id
When restarting a build job by re-submitting a job with the sourcehut
UI, the job_id changes. Update the job_id after a successful build.

This is the second time this commit is applied, now that the sourcehut
job service isn't always reporting "$BUILD_ID", but the real build id.
That had been fixed in 59f50989 ("sourcehut: properly set BPO_JOB_ID").

Fixes: #39