test: add timeout to each test
Set default timeout to 3s and a bigger timeout for two slow tests. I've
measured how long the tests take on my PC and added an extra 10 seconds
for those. Add a "_SLOW_NNs" suffix to the slow tests, so it is obvious
that these tests should take longer and they are probably not hanging.