get_repo_missing: all branches/arches in one job
Do not run a get_repo_missing job for each arch and branch, just run
one job for all of them. Put each arch/branch combination in its own
task, and also output the "repo_missing.*.json" file that will be
uploaded. Use the proper --arch parameter for
"pmbootstrap repo_missing".

Using one job for all arches and branches makes it more robust, because
it makes it unlikely that multiple results for get_repo_missing arrive
at the same time and the web server may reject one of them as it is
still processing another one. Also this makes it faster, as there is
less overhead.