jobs/build_package: pmbootstrap build --force
Use the force command, so pmbootstrap won't refuse to build a package
if there is an existing binary package with a higher version. This is
necessary when we downgraded the version in pmaports.git, which is
sometimes necessary.

Right now this is the case with postmarketos-demos and
postmarketos-splash for x86_64.

Related: it would be nice if Alpine had an epoch parameter to avoid the
downgrades: https://gitlab.alpinelinux.org/alpine/apk-tools/issues/10533