jobs/build_package: increase pmbootstrap timeout
When building locally, 900 seconds is enough to build all packages on
the old package building PC (that is building the soon to be obsoleted
binary repo). However, with sourcehut, it still runs into this timeout
when linking in some cross compilation builds (e.g. gtk+2.0-maemo).

Increase the timeout to 1500 and add it as option to
bpo/config/const.py.