Remove broken apks from abuild 3.5.0_rc1-r0
Delete apks that can't be installed with apk (#56). Reset the status of
the related db entries back to queued:
* package in the db, of which the apk was deleted
* packages depending on deleted apk, where status is failed

This patch may seem like a lot of effort for a task that could be done
manually once. But then again, doing it manually would also be a lot of
effort. Furthermore we can reuse this piece of code if we ever have a
similar situation with broken packages. Only
bpo.repo.status.is_apk_broken() needs to be adjusted then.