bpo.db: implement first database upgrade
To show, that it works as intended, here are the first two changes:

1. Package: add index "arch-branch"
This should speed up bpo.repo.status.fix_db_vs_disk() performance, to
aid with improving the overall restart speed (#57).

2. Log: add column "commit"
Can be used in the future to store the current bpo server code commit
for each restart log entry (#40).

I thought about using a migration library like alembic, but it seems
like overkill for this use case. Instead of writing down these verbose
migration files, just add very short upgrade statements to
bpo.db.migrate.upgrade(). The initial layout v0 is never modified. New
databases get bootstrapped by applying v0 and all other changes on top.

In the new upgrade function, raw SQL statements are used. For ALTER it
isn't possible to do it in another way, as noted in [1].

To create indexes, it seemed to be possible to use:
	Index("arch-branch", Package.arch, Package.branch).create(bind=engine)
But then the testsuite would not work anymore, throwing these errors
after a test case ran twice:
	sqlalchemy.exc.OperationalError: (sqlite3.OperationalError) index arch-branch already exists

Some state must still be stored in sqlalchemy. After three hours or so
of trying to figure it out, I gave up and went with the raw SQL
statement which simply works and is in my opinion not less elegant.

[1] "Altering Schmas through Migrations" at https://docs.sqlalchemy.org/en/13/core/metadata.html

Fixes: #34