ui: split bpo.db.get_all_packages_by_status()
Move code from bpo.ui.update() to new function
bpo.db.get_all_packages_by_status(). Rename .update() to .update_index()
and pass a new session argument, as well as a pkgs argument that holds
the result of the splitted function.

This is in preparation for adding badges, they require the same data.