bpo.ui: generate status image "badge.svg"
Depending on the status of the packages in the database, copy paste one
of the status svg images to /badge.svg. It reads:
* "binary repo|failed" if failed packages > 0
* "binary repo|building" if queued + building packages > 0
* "binary repo|up-to-date" otherwise

I've generated the badges with shields.io. They are a bit more complex
than I thought, so it's easier to keep just three static images around
instead of generating them dynamically.