test_zzz_slow: adjust to pkgrel bumps in pmaports
Adjust test code to version changes:
- hello-world: 1-r4 -> 1-r5
- hello-world-wrapper: 1-r2 -> 1-r3

Keep the existing versions everywhere, except in the
test_build_final_repo_with_two_pkgs test, which needs to have the same
versions as in pmaports. Because it is actually building these two
packages from the current APKBUILDs in pmaports.

In order to be able put the versions in a central place in the future,
do not duplicate the JSON payload for the get-depends job callback.
Instead, add bpo_test.override_depends_json() to create a new payload
during the test with the proper versions.

This patch makes all tests work again with current pmaports.

Fixes: #64