helpers/submit.py: move from pmaports.git
Put submit.py in this repository. It is only useful here, as changes in
submit.py always need to be syntax with the bpo code and bpo is the only
user of the script. Also, with the upcoming release branches, we would
need to sync the changes in pmaports.git across multiple branches.

To make this work, adjust the sourcehut builds and local job services to
make submit.py from build.postmarketos.org.git available during the build.