bpo/jobs: set --aports for each pmbootstrap call
The sourcehut job service only has one aports dir, and the path is set
(to $PWD/pmaports) with pmbootstrap init in the sourcehut specific
bpo_setup code (bpo.job_services.sourcehut.get_manifest()).

However, the local job service has the "real" pmaports.git dir, and a copy
of it in $PWD/pmaports, and only in the copy, the branch is switched etc.
The local job service does not run "pmbootstrap init" and configure that
copy as the dir that pmbootstrap looks into by default (otherwise it would
change the settings for the user even if not running bpo, which would be
confusing). So let's just set it explicitly with every call.