bpo.repo.status: fix obsolete_wip_package crash
Don't try to log a message with package data, if the package was not
found in the database. Log with the data parsed from the apk instead.

This can be triggered with ./bpo_local.sh, waiting until a few packages
are built, then remove bpo.db and run ./bpo_local.sh again.