bpo.repo.set_stuck: don't log to UI
Get rid of the "repo is stuck - fix failed builds to continue" messages
that are appearing over and over again once the repo hits that state.
Thanks to the big badge on top, it is super obvious if we have failing
packages, so no need to flood the log with that message. Adjust the test
accordingly.