badge: allow to ignore WIP branch errors
WIP branches that are building for the first time can be listed in the
config now, so they are ignored for the big overall status badge. We
don't want errors from these overshadow errors from brancehs that are
used in production. When the option is used, a big start (*) is
displayed next to the badge, and it links to a note that is shown in the
"failed" section.