bpo/job_services/local.py: queue multiple jobs
Multiple jobs started in parallel need to be supported, so we can get
missing packages for multiple branches.

When using the local job service, also allow the bpo server to start
multiple jobs in parallel. In the LocalJobServiceThread, queue them to
run one job after another. The sourcehut job service would really run
them in parallel, but with the local job service we are running all jobs
on the same PC, so this isn't easily possible.

With this patch, the testsuite waits for the LocalJobService to properly
terminate before moving on with the next test.

While at it, also use the same api callback for failed jobs that
sourcehut is using (instead of the fail api call that only the local job
service was using).