bpo.config.const.branches: have arches per branch
Refactor bpo.config.const.branch, so the arches and the ignore_errors
boolean are stored per branch.

Simplify the UI code by not displaying a star next to the badge if there
is a branch with ignored errors, but instead have a note printed about
that branch in the overview.