bpo.repo.status.fix: fix iterate over all arches
Use a different name for the arch variable passed to the function
(limit_arch), so it does not get overwritten by the loop iterator of the
same name. Rename the branch function argument to limit_branch for
consistency.

This bug has caused bpo.repo.status.fix to only iterate over the armv7
arch of the master branch with the current config:
- it would iterate over v20.05 branch with arches x86_64, aarch64, armv7
- while doing that, it would overwrite arch with the current iterator
- the last one would be armv7
- it would iterate over master branch with only the armv7 arch instead
  of using all available arches from the config