bpo.repo: repo with running build is incomplete
Replace count_failed_builds() with has_unfinished_builds(). Instead of
counting failed builds, return True if there is any package for the
given arch and branch in the status failed _or_building_.

This resolves a corner case, where bpo would fail to initialize if
* a build for arch/branch is still running
* all other packages of arch/branch can't be built

Add a test for that case.