sourcehut: no TLS for downloading signed WIP pkgs
Since the bpo server has been moved from the old debian stretch host to
the new alpine 3.12 one, and the server software was changed from apache
to nginx, we saw a strange segfault on builds.sr.ht when running an
aarch64 build_package job.

After hours of debugging, I found that libfetch in apk crashes when
downloading files from the WIP repo via HTTPS, if it runs in aarch64
qemu user mode emulation. I could not reproduce this locally, but I
could reproduce it on builds.sr.ht (the attaching via SSH to failed jobs
feature made reproducing and debugging feasible). More specifically, it
crashed as it called openssl's SSL_read().

It does not seem like a good use of time to go down this rabbit hole
further while backlog of other things is piling up. As workaround,
let apk download these packages from the WIP repo via HTTP instead of
HTTPS. This is still secure, because apk verifies the signature of the
APKINDEX (signed with WIP repository key), and the checksum of each .apk
package it downloads from the repository against the checksum stored in
the APKINDEX.

The sign_index job is still using HTTPS. It uses wget instead of
apk+qemu to download a single, unsigned APKINDEX from the URL.

Fixes: https://gitlab.com/postmarketOS/pmbootstrap/-/issues/1958