Build whole postmarketOS images (MR 18)
Add a new config in bpo/config/const/images.py to define which images
shall be built and in which interval. Add a new "image" table, similar
to "package". On startup, fill the image queue based on the config, and
remove obsolete database entries. Create branch:device:ui specific jobs
once all packages for a given branch are built, and a slot for starting
another build job is available. Packages are always higher prioritized
than images. Run the queue full function periodically, so we can add new
entries once the interval is reached.

Finished images are uploaded into one directory per database entry. That
directory gets a readme.html based on a new template, and it links to
the log entry that built the files in the directory.