job_services: don't use PackageStatus values (MR 18)
Add a new JobStatus enum and return these values in
bpo.job_services.*.get_status(), instead of returning the PackageStatus
values. Translate the JobStatus to PackageStatus in
bpo.helpers.job.get_status_package().

This is in preparation for building images, and changing the status of
an image based on a job status change.