job_services/local: pmbootstrap: set --aports (MR 18)
When running with the sourcehut job service, we set the pmaports dir in
"pmbootstrap init". It gets written to pmbootstrap.cfg (key "aports")
and will be used in each subsequend pmbootstrap call.

For the local job service, this isn't the case. It doesn't make sense to
change the aports dir of the user's pmbootstrap.cfg (as it affects
running pmbootstrap outside of the testsuite), so just add the correct
path to each pmbootstrap call. I discovered this, because switching the
pmaports branch didn't have any effect with the local job service: it
would still use whatever branch is checked out in the original pmaports
dir.