bpo.helpers.job: 'cancelled' -> build failure, too (MR 22)
Treat sourcehut builds status 'cancelled' as a build failure for
packages and images, too (just like the recently added translation for
'timeout').

This will make bpo retry the build until the maximum number of retries
is reached (currently 3), just like any other failed build. I've
considered making it not retry in that case, but this would involve some
refactoring and it won't matter much in practice (if you cancel a build
manually, you might as well cancel the two following builds manually, or
stop bpo altogether if it turned into a beast that is out of control).
With this change, at least bpo won't run into a RuntimeError if it sees
a 'cancelled' job.

This fixes fall-out from images related refactoring in 485c8244 ("job_services:
don't use PackageStatus values (MR 18)").