bpo/job_services/sourcehut: set extra_space = 500
pmbootstrap creates a chroot, then calculates the size of that chroot,
and creates an image file of that size to fill it with the contents.
It's a known issue that this may fail depending on the file system, but
we have the extra_space parameter to make up for that. I did a few tests
and 500 MiB seems to work fine, so use that value.

Fixes: #94