bpo.jobs.build_image: extra_space=50 for installer
The extra_space offset is only needed for the installer image. I did set
it in general for sourcehut before, but this has the problem that the
regular, non-install image also gets increased in size. So now with
extra_space=50 for the installer only, really the part where the image
was too small is adjusted, and all builds should pass again.

Related: https://builds.sr.ht/~ollieparanoid/job/474398