style.css: make bottom bar brighter (MR 26)
Change to same color as the navigation bar in the header. I'm about to
use the same style.css for the new bpo directory listing. There the
actual content hasn't much height and so the footer is immediately
visible. Make it brighter so it's not distracting from the actual
content.