config: remove v20.05 (MR 31)
Stop building packages for v20.05. By disabling this, we save some time
while scanning for packages that need to be built (done on startup).

Also due to bug #69, whenever deleting a package, the APKINDEXes of all
branches get regenerated (not just the branch where it was deleted). So
until this is fixed, it's at least nice to not have too many branches
where they get regenerated.